% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cook_rcbd.R
\name{cook_rcbd}
\alias{cook_rcbd}
\title{วิเคราะห์แผนการทดลองแบบ RCBD (หุงข้าว RCBD)}
\usage{
cook_rcbd(data, outcome, treatment, block)
}
\arguments{
\item{data}{data.frame ที่มีข้อมูล}

\item{outcome}{ชื่อคอลัมน์ของผลผลิต (เช่น yield)}

\item{treatment}{ชื่อคอลัมน์ของกรรมวิธี (เช่น variety)}

\item{block}{ชื่อคอลัมน์ของ Block (เช่น replication)}
}
\value{
รายการผลลัพธ์ ANOVA
}
\description{
วิเคราะห์แผนการทดลองแบบ RCBD (หุงข้าว RCBD)
}
