% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_with_sig.R
\name{top_with_sig}
\alias{top_with_sig}
\title{ตรวจสอบความแตกต่างและใส่ตัวอักษร (โรยหน้าด้วยสถิติ)}
\usage{
top_with_sig(model, treatment)
}
\arguments{
\item{model}{ผลลัพธ์จาก cook_rcbd}

\item{treatment}{ชื่อคอลัมน์ของกรรมวิธี}
}
\description{
ตรวจสอบความแตกต่างและใส่ตัวอักษร (โรยหน้าด้วยสถิติ)
}
