% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wash_rice.R
\name{wash_rice}
\alias{wash_rice}
\title{ทำความสะอาดข้อมูลเบื้องต้น (ล้างข้าว)}
\usage{
wash_rice(data)
}
\arguments{
\item{data}{data.frame ที่ต้องการทำความสะอาด}
}
\value{
data.frame ที่ลบแถวที่มี NA ออกแล้ว
}
\description{
ลบแถวที่มีค่า NA ออกจากข้อมูลก่อนนำไปวิเคราะห์ด้วย [cook_rcbd()]
}
\examples{
df <- data.frame(x = c(1, 2, NA, 4), y = c("a", "b", "c", NA))
wash_rice(df)

}
